Use a single websocket per browser
