+++ .gitignore
@@ -1,0 +1,3 @@
+.env
+node_modules
+public/app.js
